refactor(tests): move auth tests from source to tests/core/auth/
Move SCRAM-SHA-256 and MD5 tests to integration test directory.
Make tested helper functions pub for external test access.
