+++ tests/core/auth/mod.rs
@@ -1,0 +1,2 @@
+mod md5;
+mod scram;
